issue-40: usar componente dinámico para que pueda imprimir el PDF
